Add ability to change foliage far distance.
The FoliageLevelManager can now adjust the maximum distance that the foliage can
appear at. This is done by retrieving the GeometryPageManagers that manage each
detail level and then setting their far distances. The far distance is set for the
furthest level and then the rest are offset by using the far distance factor desired.
A minimum needs to be set on the far distance.

Names for functions were changed from update (eg. updateDensity) to set (eg.
setDensity) to better match standards.