Add a mimimum distance factor.
Earlier the foliage distance could also be set to 0, which is the same as changing the
foliage density to 0. Thus when density was updated it would have to wait for the far
distance to update too before it would take effect. If they happened slightly out of
sync the density would only update on the next reload. With a minimum far distance
this effect should be reduced.