Remove different desired fps slider and add better frame limiting detection.
- When the average fps is in very close range of desired fps, the frame limting bool is
checked to see if the frame rate is being limited. This would imply that there is
headroom to improve the graphics detail. The system then requests a change which
should theoretically drop the fps by 8. If the fps doesn't drop, it will try it again,
untill it does drop, at which point the system will recover a part of the fps.

- The boolean is accessed through the main loop controller.

- Also remove leftover int to float casting during comparison in foliage level
manager.